// src/api/models/experienceResDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { ExperienceResDTOHopeJob } from './experienceResDTOHopeJob';

export interface ExperienceResDTO {
  hopeJob: ExperienceResDTOHopeJob;
  id: number;
  name: string;
  createdAt: string;
}